requirements.txt
wikipedia
pydocumentdb
spotipy
azure-keyvault>=0.3.6
azure-mgmt-keyvault